Detached PTX(pendint transaction analysis)
