+++ npms/skale-ima/index.js
@@ -43,6 +43,8 @@ cc.enable( false );
 log.addStdout();
 // log.add( strFilePath, nMaxSizeBeforeRotation, nMaxFilesCount ); // example: log output to file
 
+const ptx = require( "./ptx.js" );
+
 const owaspUtils = require( "../skale-owasp/owasp-util.js" );
 const { getWeb3FromURL } = require( "../../agent/cli.js" );
 
@@ -854,7 +856,7 @@ async function do_oracle_gas_price_setup(
                 } );
                 print_gas_usage_report_from_array( "(intermediate result) ORACLE GAS PRICE SETUP ", jarrReceipts );
                 // if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-                //     await async_pending_tx_complete( details, w3_schain, w3_main_net, chain_id_schain, chain_id_mainnet, "" + joReceipt.transactionHash );
+                //     await ptx.complete( details, w3_schain, w3_main_net, chain_id_schain, chain_id_mainnet, "" + joReceipt.transactionHash );
             }
             save_transfer_success( "oracle" );
         } );
@@ -4769,279 +4771,6 @@ function parseIntOrHex( s ) {
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-async function async_pending_tx_start( details, w3, w3_opposite, chain_id, chain_id_opposite, txHash ) {
-    const strLogPrefix = "";
-    try {
-        if( chain_id == "Mainnet" ) {
-            details.write( strLogPrefix + cc.debug( "Reporting pending transaction " ) + cc.notice( txHash ) + + cc.debug( " start from " ) + cc.u( owaspUtils.w3_2_url( w3 ) ) + cc.debug( "..." ) + "\n" );
-            const strNodeURL = owaspUtils.w3_2_url( w3_opposite );
-            details.write( strLogPrefix + cc.debug( "Will report pending work cache to " ) + cc.u( strNodeURL ) + cc.debug( "..." ) + "\n" );
-            const rpcCallOpts = null;
-            await rpcCall.create( strNodeURL, rpcCallOpts, async function( joCall, err ) {
-                if( err ) {
-                    const s = cc.fatal( "PENDING WORK START ERROR:" ) + cc.error( " JSON RPC call to S-Chain node failed" ) + "\n";
-                    if( verbose_get() >= RV_VERBOSE.error )
-                        log.write( s );
-                    details.write( s );
-                    if( joCall )
-                        await joCall.disconnect();
-                    return;
-                }
-                const joIn = {
-                    method: "skale_imaTxnInsert",
-                    params: {
-                        hash: "" + txHash
-                    }
-                };
-                details.write( cc.debug( "Starting pending work with " ) + cc.j( joIn ) + "\n" );
-                await joCall.call( joIn, async function( joIn, joOut, err ) {
-                    if( err ) {
-                        const s = cc.fatal( "PENDING WORK START ERROR:" ) + cc.error( " JSON RPC call to S-Chain node, error: " ) + cc.warning( err ) + "\n";
-                        if( verbose_get() >= RV_VERBOSE.error )
-                            log.write( s );
-                        details.write( s );
-                        await joCall.disconnect();
-                        return;
-                    }
-                    details.write( cc.debug( "Pending work start result is: " ) + cc.j( joOut ) + "\n" );
-                    if( joOut && "result" in joOut && "success" in joOut.result ) {
-                        if( joOut.result.success ) {
-                            details.write( strLogPrefix + cc.success( "Success, pending work start reported" ) + "\n" );
-                            await joCall.disconnect();
-                            return;
-                        } else {
-                            details.write( strLogPrefix + cc.warning( "Pending work start was not reported with success" ) + "\n" );
-                            await joCall.disconnect();
-                            return;
-                        }
-                    } else {
-                        const s = cc.fatal( "PENDING WORK START ERROR:" ) + cc.error( " JSON RPC call to S-Chain node, returned bad answer: " ) + cc.j( joOut ) + "\n";
-                        if( verbose_get() >= RV_VERBOSE.error )
-                            log.write( s );
-                        details.write( s );
-                        await joCall.disconnect();
-                        return;
-                    }
-                } );
-            } );
-        }
-    } catch ( err ) {
-        const s =
-            strLogPrefix + cc.error( "PENDING WORK START ERROR: API call error from " ) + cc.u( owaspUtils.w3_2_url( w3 ) ) +
-            cc.error( ": " ) + cc.error( err ) +
-            "\n";
-        if( verbose_get() >= RV_VERBOSE.error )
-            log.write( s );
-        details.write( s );
-    }
-}
-
-async function async_pending_tx_complete( details, w3, w3_opposite, chain_id, chain_id_opposite, txHash ) {
-    const strLogPrefix = "";
-    try {
-        if( chain_id == "Mainnet" ) {
-            details.write( strLogPrefix + cc.debug( "Reporting pending transaction " ) + cc.notice( txHash ) + + cc.debug( " completion from " ) + cc.u( owaspUtils.w3_2_url( w3 ) ) + cc.debug( "..." ) + "\n" );
-            const strNodeURL = owaspUtils.w3_2_url( w3_opposite );
-            details.write( strLogPrefix + cc.debug( "Will report pending work cache to " ) + cc.u( strNodeURL ) + cc.debug( "..." ) + "\n" );
-            const rpcCallOpts = null;
-            await rpcCall.create( strNodeURL, rpcCallOpts, async function( joCall, err ) {
-                if( err ) {
-                    const s = cc.fatal( "PENDING WORK COMPLETE ERROR:" ) + cc.error( " JSON RPC call to S-Chain node failed" ) + "\n";
-                    if( verbose_get() >= RV_VERBOSE.error )
-                        log.write( s );
-                    details.write( s );
-                    if( joCall )
-                        await joCall.disconnect();
-                    return;
-                }
-                const joIn = {
-                    method: "skale_imaTxnErase",
-                    params: {
-                        hash: "" + txHash
-                    }
-                };
-                details.write( cc.debug( "Completing pending work with " ) + cc.j( joIn ) + "\n" );
-                await joCall.call( joIn, async function( joIn, joOut, err ) {
-                    if( err ) {
-                        const s = cc.fatal( "PENDING WORK COMPLETE ERROR:" ) + cc.error( " JSON RPC call to S-Chain node, error: " ) + cc.warning( err ) + "'n";
-                        if( verbose_get() >= RV_VERBOSE.error )
-                            log.write( s );
-                        details.write( s );
-                        await joCall.disconnect();
-                        return;
-                    }
-                    details.write( cc.debug( "Pending work complete result is: " ) + cc.j( joOut ) + "\n" );
-                    if( joOut && "result" in joOut && "success" in joOut.result ) {
-                        if( joOut.result.success ) {
-                            details.write( strLogPrefix + cc.success( "Success, pending work complete reported" ) + "\n" );
-                            await joCall.disconnect();
-                            return;
-                        } else {
-                            await joCall.disconnect();
-                            details.write( strLogPrefix + cc.warning( "Pending work complete was not reported with success" ) + "\n" );
-                            return;
-                        }
-                    } else {
-                        const s = cc.fatal( "PENDING WORK COMPLETE ERROR:" ) + cc.error( " JSON RPC call to S-Chain node, returned bad answer: " ) + cc.j( joOut ) + "\n";
-                        if( verbose_get() >= RV_VERBOSE.error )
-                            log.write( s );
-                        details.write( s );
-                        await joCall.disconnect();
-                        return;
-                    }
-                } );
-            } );
-
-        }
-    } catch ( err ) {
-        const s =
-            strLogPrefix + cc.error( "PENDING WORK COMPLETE ERROR: API call error from " ) + cc.u( owaspUtils.w3_2_url( w3 ) ) +
-            cc.error( ": " ) + cc.error( err ) +
-            "\n";
-        if( verbose_get() >= RV_VERBOSE.error )
-            log.write( s );
-        details.write( s );
-    }
-}
-
-// function isIterable( value ) {
-//     return Symbol.iterator in Object( value );
-// }
-
-async function async_pending_tx_scanner( details, w3, w3_opposite, chain_id, chain_id_opposite, cb ) {
-    cb = cb || function( tx ) { };
-    const strLogPrefix = "";
-    try {
-        details.write( strLogPrefix + cc.debug( "Scanning pending transactions from " ) + cc.u( owaspUtils.w3_2_url( w3 ) ) + cc.debug( "..." ) + "\n" );
-        if( chain_id == "Mainnet" ) {
-            const strNodeURL = owaspUtils.w3_2_url( w3_opposite );
-            details.write( strLogPrefix + cc.debug( "Using pending work cache from " ) + cc.u( strNodeURL ) + cc.debug( "..." ) + "\n" );
-            let havePendingWorkInfo = false;
-            const rpcCallOpts = null;
-            await rpcCall.create( strNodeURL, rpcCallOpts, async function( joCall, err ) {
-                if( err ) {
-                    const s = cc.fatal( "PENDING WORK CACHE ERROR:" ) + cc.error( " JSON RPC call to S-Chain node failed" ) + "\n";
-                    if( verbose_get() >= RV_VERBOSE.error )
-                        log.write( s );
-                    details.write( s );
-                    return;
-                }
-                const joIn = {
-                    method: "skale_imaTxnListAll",
-                    params: {}
-                };
-                details.write( cc.debug( "Calling pending work cache with " ) + cc.j( joIn ) + "\n" );
-                await joCall.call( joIn, async function( joIn, joOut, err ) {
-                    if( err ) {
-                        havePendingWorkInfo = true;
-                        const s = cc.fatal( "PENDING WORK CACHE ERROR:" ) + cc.error( " JSON RPC call to S-Chain node, error: " ) + cc.warning( err );
-                        if( verbose_get() >= RV_VERBOSE.error )
-                            log.write( s );
-                        details.write( s );
-                        return;
-                    }
-                    details.write( cc.debug( "Pending work cache result is: " ) + cc.j( joOut ) + "\n" );
-                    if( joOut && "result" in joOut && "success" in joOut.result ) {
-                        if( joOut.result.success && "allTrackedTXNs" in joOut.result && joOut.result.allTrackedTXNs.length > 0 ) {
-                            details.write( strLogPrefix + cc.debug( "Got " ) + cc.j( joOut.result.allTrackedTXNs.length ) + cc.debug( " pending transaction(s)" ) + "\n" );
-                            let cntReviewedTNXs = 0;
-                            for( const joTXE of joOut.result.allTrackedTXNs ) {
-                                ++ cntReviewedTNXs;
-                                const joTX = {
-                                    hash: "" + joTXE.hash,
-                                    to: "holder.value.MessageProxy"
-                                };
-                                const isContinue = cb( joTX );
-                                if( ! isContinue )
-                                    break;
-                            }
-                            details.write( strLogPrefix + cc.debug( "Reviewed " ) + cc.info( cntReviewedTNXs ) + cc.debug( " pending transaction(s)" ) + "\n" );
-                            havePendingWorkInfo = true;
-                            return;
-                        } else {
-                            details.write( strLogPrefix + cc.debug( "Reviewed " ) + cc.info( 0 ) + cc.debug( " pending transaction(s)" ) + "\n" );
-                            havePendingWorkInfo = true;
-                            return;
-                        }
-                    } else {
-                        const s = cc.fatal( "PENDING WORK CACHE ERROR:" ) + cc.error( " JSON RPC call to S-Chain node, returned bad answer: " ) + cc.j( joOut );
-                        if( verbose_get() >= RV_VERBOSE.error )
-                            log.write( s );
-                        details.write( s );
-                        havePendingWorkInfo = true;
-                        return;
-                    }
-                } );
-            } );
-            let nWaitAttempt = 0;
-            while( ! havePendingWorkInfo ) {
-                await sleep( 3000 );
-                ++ nWaitAttempt;
-                if( nWaitAttempt >= 10 ) {
-                    havePendingWorkInfo = true; // workaround for es-lint
-                    const s = cc.fatal( "PENDING WORK CACHE ERROR:" ) + cc.error( " Wait timeout" ) + "\n";
-                    if( verbose_get() >= RV_VERBOSE.error )
-                        log.write( s );
-                    details.write( s );
-                    break;
-                }
-            }
-        } else { // if( chain_id == "Mainnet" )
-            const mapTXs = {};
-            let tx_idx = 0;
-            while( true ) {
-                let have_next_tx = false;
-                await w3.eth.getTransactionFromBlock( "pending", tx_idx, function( err, rslt ) {
-                    try {
-                        if( err )
-                            return;
-                        if( ! rslt )
-                            return;
-                        if( typeof rslt != "object" )
-                            return;
-                        have_next_tx = true;
-                        const hash = "" + rslt.hash;
-                        if( hash in mapTXs )
-                            return;
-                        mapTXs[hash] = rslt;
-                        // details.write( strLogPrefix + cc.debug( "Got pending transaction: " ) + cc.j( rslt ) + "\n" );
-                        const isContinue = cb( rslt );
-                        if( ! isContinue ) {
-                            have_next_tx = false;
-                            return;
-                        }
-                    } catch ( err ) {
-                        if( verbose_get() >= RV_VERBOSE.error ) {
-                            const s =
-                                strLogPrefix + cc.error( "PENDING TRANSACTIONS ENUMERATION HANDLER ERROR: from " ) + cc.u( owaspUtils.w3_2_url( w3 ) ) +
-                                cc.error( ": " ) + cc.error( err ) +
-                                "\n";
-                            if( verbose_get() >= RV_VERBOSE.error )
-                                log.write( s );
-                            details.write( s );
-                        }
-                    }
-                } );
-                if( ! have_next_tx )
-                    break;
-                ++ tx_idx;
-            }
-            details.write( strLogPrefix + cc.debug( "Got " ) + cc.j( Object.keys( mapTXs ).length ) + cc.debug( " pending transaction(s)" ) + "\n" );
-        } // else from if( chain_id == "Mainnet" )
-    } catch ( err ) {
-        const s =
-            strLogPrefix + cc.error( "PENDING TRANSACTIONS SCAN ERROR: API call error from " ) + cc.u( owaspUtils.w3_2_url( w3 ) ) +
-            cc.error( ": " ) + cc.error( err ) +
-            "\n";
-        if( verbose_get() >= RV_VERBOSE.error )
-            log.write( s );
-        details.write( s );
-    }
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
 async function find_out_reference_log_record( details, w3, jo_message_proxy, nBlockId, nMessageNumberToFind, isVerbose ) {
     const strLogPrefix = "";
     const bnMessageNumberToFind = w3.utils.toBN( nMessageNumberToFind.toString() );
@@ -5498,7 +5227,7 @@ async function do_transfer(
                 let wasIgnoredPTX = false;
                 try {
                     const strShortMessageProxyAddressToCompareWith = owaspUtils.remove_starting_0x( jo_message_proxy_dst.options.address ).toLowerCase();
-                    await async_pending_tx_scanner( details, w3_dst, w3_src, chain_id_dst, chain_id_src, function( joTX ) {
+                    await ptx.scanner( details, w3_dst, w3_src, chain_id_dst, chain_id_src, function( joTX ) {
                         if( "to" in joTX ) {
                             if( joTX.to == "holder.value.MessageProxy" ) {
                                 joFoundPendingTX = joTX;
@@ -5525,7 +5254,7 @@ async function do_transfer(
                             await sleep( optsPendingTxAnalysis.nTimeoutSecondsBeforeSecondAttempt * 1000 );
                             //
                             joFoundPendingTX = null;
-                            await async_pending_tx_scanner( w3_dst, w3_src, chain_id_dst, chain_id_src, function( joTX ) {
+                            await ptx.scanner( w3_dst, w3_src, chain_id_dst, chain_id_src, function( joTX ) {
                                 if( "to" in joTX ) {
                                     if( joTX.to == "holder.value.MessageProxy" ) {
                                         joFoundPendingTX = joTX;
@@ -5913,10 +5642,10 @@ async function do_transfer(
                         let joReceipt = null;
                         if( joPostIncomingMessagesSR.joACI.isAutoSend ) {
                             //
-                            // NOTICE: async_pending_tx_start()/ async_pending_tx_complete() must be called only in do_transfer()
+                            // NOTICE: ptx.start()/ ptx.complete() must be called only in do_transfer()
                             //
                             if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-                                await async_pending_tx_start( detailsB, w3_dst, w3_src, chain_id_dst, chain_id_src, "" + joPostIncomingMessagesSR.txHashSent );
+                                await ptx.start( detailsB, w3_dst, w3_src, chain_id_dst, chain_id_src, "" + joPostIncomingMessagesSR.txHashSent );
                             joReceipt = await get_web3_transactionReceipt( detailsB, 10, w3_dst, joPostIncomingMessagesSR.txHashSent );
                         } else {
                             const serializedTx_postIncomingMessages = tx_postIncomingMessages.serialize();
@@ -5933,10 +5662,10 @@ async function do_transfer(
                             } );
                             print_gas_usage_report_from_array( "(intermediate result) TRANSFER " + chain_id_src + " -> " + chain_id_dst, jarrReceipts );
                             //
-                            // NOTICE: async_pending_tx_start()/ async_pending_tx_complete() must be called only in do_transfer()
+                            // NOTICE: ptx.start()/ ptx.complete() must be called only in do_transfer()
                             //
                             if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-                                await async_pending_tx_complete( detailsB, w3_dst, w3_src, chain_id_dst, chain_id_src, "" + joReceipt.transactionHash );
+                                await ptx.complete( detailsB, w3_dst, w3_src, chain_id_dst, chain_id_src, "" + joReceipt.transactionHash );
                         }
                         cntProcessed += cntAccumulatedForBlock;
                         //
@@ -6430,7 +6159,7 @@ async function mintERC20(
         let joReceipt = null;
         if( joSR.joACI.isAutoSend ) {
             // if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-            //     await async_pending_tx_start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
+            //     await ptx.start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
             joReceipt = await get_web3_transactionReceipt( details, 10, w3, joSR.txHashSent );
         } else {
             const serializedTx_mint = tx_mint.serialize();
@@ -6526,7 +6255,7 @@ async function mintERC721(
         let joReceipt = null;
         if( joSR.joACI.isAutoSend ) {
             // if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-            //     await async_pending_tx_start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
+            //     await ptx.start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
             joReceipt = await get_web3_transactionReceipt( details, 10, w3, joSR.txHashSent );
         } else {
             const serializedTx_mint = tx_mint.serialize();
@@ -6625,7 +6354,7 @@ async function mintERC1155(
         let joReceipt = null;
         if( joSR.joACI.isAutoSend ) {
             // if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-            //     await async_pending_tx_start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
+            //     await ptx.start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
             joReceipt = await get_web3_transactionReceipt( details, 10, w3, joSR.txHashSent );
         } else {
             const serializedTx_mint = tx_mint.serialize();
@@ -6721,7 +6450,7 @@ async function burnERC20(
         let joReceipt = null;
         if( joSR.joACI.isAutoSend ) {
             // if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-            //     await async_pending_tx_start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
+            //     await ptx.start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
             joReceipt = await get_web3_transactionReceipt( details, 10, w3, joSR.txHashSent );
         } else {
             const serializedTx_burn = tx_burn.serialize();
@@ -6817,7 +6546,7 @@ async function burnERC721(
         let joReceipt = null;
         if( joSR.joACI.isAutoSend ) {
             // if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-            //     await async_pending_tx_start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
+            //     await ptx.start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
             joReceipt = await get_web3_transactionReceipt( details, 10, w3, joSR.txHashSent );
         } else {
             const serializedTx_burn = tx_burn.serialize();
@@ -6915,7 +6644,7 @@ async function burnERC1155(
         let joReceipt = null;
         if( joSR.joACI.isAutoSend ) {
             // if( optsPendingTxAnalysis && "isEnabled" in optsPendingTxAnalysis && optsPendingTxAnalysis.isEnabled )
-            //     await async_pending_tx_start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
+            //     await ptx.start( details, w3, w3, cid, cid, "" + joSR.txHashSent );
             joReceipt = await get_web3_transactionReceipt( details, 10, w3, joSR.txHashSent );
         } else {
             const serializedTx_burn = tx_burn.serialize();
@@ -6949,6 +6678,7 @@ module.exports.longSeparator = g_mtaStrLongSeparator;
 module.exports.noop = noop;
 module.exports.cc = cc;
 module.exports.log = log;
+module.exports.ptx = ptx;
 module.exports.sleep = sleep;
 module.exports.owaspUtils = owaspUtils;
 module.exports.w3mod = w3mod;
